+++ src/MessageResponseStatus.ts
@@ -1,3 +1,4 @@
 export enum MessageResponseStatus {
   Success = 0,
+  SizeLimitExceeded = 4,
 }